infrastructure: Remove options which depend on local setup
These were automatically added by `pulumi new`, but a developer might
very well choose to not use pip, and / or put their virtual environments
elsewhere.

In particular, this caused issues for running end-to-end tests as pulumi
commands would look for a venv in the `infrastructure/` directory...
which you might not have if you're just running tests from the `tests/`
directory.